requirements.txt
streamlit
pdfplumber
nltk
sentence-transformers
scikit-learn
pandas
numpy
ctransformers
python-docx
beautifulsoup4
requests
youtube-transcript-api
graphviz